requirements.txt
django
djangorestframework
python_version >= '3.8'